Updated control flow and UI interactions in CWorkorderPartsDialog
This commit includes several updates to the control flow and user interface interactions in the `CWorkorderPartsDialog` class. The `PreTranslateMessage` function has been updated in several files to return `CDialogEx::PreTranslateMessage(pMsg)` when the `if` condition is not met. A new `PreTranslateMessage` function has been added to `CWorkorderPartsDialog.cpp` to handle `WM_KEYDOWN` messages. The `OnInitDialog`, `OnEnChangeWorkorderAddParts`, `OnNMDoubleClickWorkorderStockPartsList`, `OnNMClickWorkorderAddedPartsList`, `OnBnClickedWorkorderAddPart`, `OnBnClickedWorkorderChange`, and `OnBnClickedWorkorderDeleteAddedPart` functions have been updated to improve focus management and button interaction. A new `SetCustomFocusButton` function has been added to set the text color and focus of a button. The `CWorkorderPartsDialog` class has been updated to change the type of several button members from `CButton` to `CMFCButton` and add the `PreTranslateMessage` function to the class declaration. The `OnNMDoubleClickWorkorderViewExisting` function in `CWorkorderView.cpp` has been updated to call the `SetCustomFocusButton` function.