Updated SQL handling and conditions in CWorkorderView functions
This commit includes several updates to the `CWorkorderView.cpp` and `CWorkorderView.h` files. The update date in the comments has been changed to 26-05-2024. The condition for executing the contribution dialog in `CWorkorderView::OnBnClickedWorkorderViewClose()` has been simplified to execute regardless of the contribution amount. The SQL query execution for inserting into the `CONTRIBUTION` and `INVOICE` tables has been restructured for better efficiency. The `CWorkorderView::PerformWorkorderUpdate()` function now includes an additional SQL query to update the `ASSET` table if the `m_chbWorkorderAssetDisposed` checkbox is checked. The SQL connection is now properly closed after the queries are executed in both functions. Lastly, the `strQuery.ReleaseBuffer()` function is now called after each SQL query execution to free up memory.