Added new font and updated UI elements in various files
In this commit, several changes were made across multiple files. In `CPrintWorkorder.cpp`, a new font `fontPlainBold` was added to the `PrintCombi` function for printing customer workorder details. The status bar text in `CWorkorderView.cpp` was updated to reflect the correct action. The MessageBox function in `RepairCafeCureApp.cpp` was updated to use the main window handle as the parent window. The version of `CAboutDlg` class was updated to 1.0.0.1 (alpha). Two new resources were added to `RepairCafeCureApp.vcxproj` and `RepairCafeCureApp.vcxproj.filters`. New control values were added in `Resource.h` and a new bitmap file `bitmap1.bmp` was added.