Added CReportTaxView class and updated resources for tax report generation
A new class, CReportTaxView, was added for generating a tax contribution report. This class allows users to select a period, generate a report, and print it. The class was implemented with various methods for data exchange, initialization, message translation, printing preparation, and setting the printer orientation. New identifiers were added in Resource.h and new image resources were added in RepairCafeCureApp.vcxproj.filters. The bitmap1.bmp resource was replaced with LogoAVBIcon.ico. A new category and panel were added in ribbon.mfcribbon-ms. A copyright notice and detailed comment block were added at the beginning of the file.